+++ frontend/scripts/componentSchemas/run.ts
@@ -1,28 +1,7 @@
-import axios from 'axios';
-import jsonpointer from 'jsonpointer';
-
-const versionSettings = {
-  v3: {
-    layoutSchemaUrl: 'https://altinncdn.no/schemas/json/layout/layout.schema.v1.json',
-    componentSchemaPath: '../../packages/ux-editor-v3/src/testing/schemas/json/component',
-  },
-  v4: {
-    layoutSchemaUrl:
-      'https://altinncdn.no/toolkits/altinn-app-frontend/4.0.0-rc2/schemas/json/layout/layout.schema.v1.json',
-    componentSchemaPath: '../../packages/ux-editor/src/testing/schemas/json/component',
-  },
-};
-const validVersions = ['v3', 'v4'] as const;
-type AppFrontendVersion = (typeof validVersions)[number];
-
-const isValidVersion = (version: string): boolean => {
-  return validVersions.includes(version as AppFrontendVersion);
-};
-
-const getLayoutSchema = async (version?: AppFrontendVersion) => {
-  const response = await axios.get(versionSettings[version || 'v4'].layoutSchemaUrl);
-  return response?.data;
-};
+import { expandAllOf, expandAnyOf, expandRefsInProperties, verifySchema } from './schemaUtils';
+import type { AppFrontendVersion } from './version';
+import { isValidVersion, versionSettings } from './version';
+import { getLayoutSchema } from './api';
 
 const writeToFile = (name: string, data: any, version: AppFrontendVersion) => {
   const path = require('path');
@@ -35,63 +14,6 @@ const writeToFile = (name: string, data: any, version: AppFrontendVersion) => {
     if (err) return console.log(err);
     console.log(`Wrote ${fileName}`);
   });
-};
-
-const expandAllOf = (schema: any, layoutSchema: any, componentNode: boolean = true) => {
-  const expandedSchema: any = {};
-  const allOfList = componentNode ? schema.allOf?.slice(0, -1) : schema.allOf;
-  allOfList.forEach((item: any) => {
-    if (item.$ref) {
-      let ref = jsonpointer.get(layoutSchema, item.$ref.replace('#/', '/'));
-
-      if (ref.allOf) {
-        ref = expandAllOf(ref, layoutSchema, false);
-      }
-      expandedSchema.properties = { ...expandedSchema.properties, ...ref.properties };
-    } else if (item.properties) {
-      expandedSchema.properties = { ...expandedSchema.properties, ...item.properties };
-    }
-  });
-  expandedSchema.required = schema.allOf?.[schema.allOf.length - 1].required;
-  return expandedSchema;
-};
-
-const expandAnyOf = (schema: any, layoutSchema: any) => {
-  const anyOf = [];
-  schema.anyOf?.forEach((item: any) => {
-    if (item.$ref) {
-      let ref = jsonpointer.get(layoutSchema, item.$ref.replace('#/', '/'));
-      if (ref.allOf) {
-        ref = expandAllOf(ref, layoutSchema, false);
-      }
-      anyOf.push(ref);
-    } else if (item.properties) {
-      anyOf.push(item);
-    }
-  });
-  return anyOf;
-};
-const expandRef = (ref: string, layoutSchema: any) => {
-  return jsonpointer.get(layoutSchema, ref.replace('#/', '/'));
-};
-
-const ensureStringTypeWithEnums = (schema: any) => {
-  if (schema.enum) {
-    schema.type = 'string';
-  }
-};
-
-const verifySchema = (schema: any, expectedProperties: string[]) => {
-  const schemaProperties = Object.keys(schema.properties);
-  const missingProperties = expectedProperties.filter(
-    (property) => !schemaProperties.includes(property),
-  );
-  if (missingProperties.length > 0) {
-    console.log(`Missing properties: ${missingProperties.join(', ')}`);
-    return false;
-  }
-
-  return true;
 };
 
 const generateComponentSchema = (name: string, layoutSchema: any) => {
@@ -124,30 +46,6 @@ const generateComponentSchema = (name: string, layoutSchema: any) => {
   return schema;
 };
 
-const expandRefsInProperties = (properties: any, layoutSchema: any) => {
-  const expandedProperties = { ...properties };
-  for (const property in properties) {
-    if (
-      expandedProperties[property].$ref &&
-      !expandedProperties[property].$ref.startsWith('expression.schema')
-    ) {
-      expandedProperties[property] = expandRef(expandedProperties[property].$ref, layoutSchema);
-    }
-    if (expandedProperties[property].items?.$ref) {
-      expandedProperties[property].items = expandRef(
-        expandedProperties[property].items.$ref,
-        layoutSchema,
-      );
-    }
-    if (expandedProperties[property].allOf) {
-      expandedProperties[property] = expandAllOf(expandedProperties[property], layoutSchema);
-    }
-    ensureStringTypeWithEnums(expandedProperties[property]);
-  }
-
-  return expandedProperties;
-};
-
 const run = async () => {
   let version: string = process.argv.length > 2 ? process.argv[2] : '';
   if (!isValidVersion(version)) {
@@ -164,7 +62,11 @@ const run = async () => {
       componentName === 'AddressComponent' ? 'Address' : componentName,
       layoutSchema,
     );
-    writeToFile(componentName, schema, version as AppFrontendVersion);
+    writeToFile(
+      componentName === 'AddressComponent' ? 'Address' : componentName,
+      schema,
+      version as AppFrontendVersion,
+    );
   });
 };
